requirements.txt
apache-airflow==2.8.0
apache-airflow-providers-databricks
apache-airflow-providers-amazon
apache-airflow-providers-snowflake
databricks-cli
snowflake-connector-python
pytest
boto3